Fix routers for user functions
