--- server/controller/signOut.js
@@ -1,22 +1,0 @@
-module.exports = {
-
-	signOut: async (req, res) => {
-        
-        // TODO: 로그아웃을 하고 세션을 종료한다.
-        // status: 200
-        // {
-        //     "message": "Logout success"
-        // }
-        // status:401
-        // {
-        //     "message": "Unauthorized user"
-        // }
-        // status:404
-        // {
-        // "message": "Not found"
-        // }
-        
-        return res.send('signOut');
-
-	}
-};
