+++ server/controller/users.js
@@ -36,5 +36,50 @@ module.exports = {
         
         return res.send('signUp');
 
+	},
+    
+    signOut: async (req, res) => {
+        
+        // TODO: 로그아웃을 하고 세션을 종료한다.
+        // status: 200
+        // {
+        //     "message": "Logout success"
+        // }
+        // status:401
+        // {
+        //     "message": "Unauthorized user"
+        // }
+        // status:404
+        // {
+        // "message": "Not found"
+        // }
+        
+        return res.send('signOut');
+
+	},
+
+    signIn: async (req, res) => {
+        
+        // TODO: 사용자의 email, password를 조회해 로그인한다. 
+        // req.body
+        // {
+        //     "user_email": "example@gmail.com",
+        //     "user_password": "string",
+        // }
+        // status: 200
+        // {
+        //     "message": "Login success"
+        // }
+        // status:401
+        // {
+        //     "message": "Invalid password"
+        // }
+        // status:404
+        // {
+        //     "message": "Invalid user"
+        // }
+        
+        return res.send('signIn');
+
 	}
 };
