--- server/router/signInRouter.js
@@ -1,7 +1,0 @@
-const { signIn } = require('../controller/signIn');
-const express = require('express');
-const router = express.Router();
-
-router.post('/', signIn);
-
-module.exports = router;
