--- server/router/signOutRouter.js
@@ -1,7 +1,0 @@
-const { signOut } = require('../controller/signOut');
-const express = require('express');
-const router = express.Router();
-
-router.post('/', signOut);
-
-module.exports = router;
